Move StreamGobbler to shell package
